requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
requests==2.31.0